+++ stellar_sdk/memo.py
@@ -178,10 +178,9 @@ class HashMemo(Memo):
     def __init__(self, memo_hash: Union[bytes, str]) -> None:
         memo_hash = hex_to_bytes(memo_hash)
         length = len(memo_hash)
-        # TODO: fix
-        if length > 32:
+        if length != 32:
             raise MemoInvalidException(
-                "HashMemo can contain 32 bytes, got {:d} bytes".format(length)
+                "The length of HashMemo should be 32 bytes, got {:d} bytes.".format(length)
             )
 
         self.memo_hash: bytes = memo_hash
@@ -217,10 +216,11 @@ class ReturnHashMemo(Memo):
     def __init__(self, memo_return: bytes) -> None:
         memo_return = hex_to_bytes(memo_return)
         length = len(memo_return)
-        # TODO: fix
-        if length > 32:
+        if length != 32:
             raise MemoInvalidException(
-                "ReturnHashMemo can contain 32 bytes, got {:d} bytes".format(length)
+                "The length of ReturnHashMemo should be 32 bytes, got {:d} bytes.".format(
+                    length
+                )
             )
 
         self.memo_return: bytes = memo_return